+++ modules/i18n/src/models/namespaces/flows-ns.ts
@@ -36,4 +36,7 @@ export interface flowsNS {
     passwordRecovery: {
         name: string;
     };
+    askPassword: {
+        breadcrumb: string;
+    };
 }